Move files from src to frontend
