+++ frontend/src/app/features/new-dashboard/link-to-parent/link-to-parent.component.html
