+++ frontend/src/app/features/new-dashboard/parent-home-tab/parent-home-tab.component.scss
