+++ frontend/src/app/features/new-dashboard/remove-link-to-parent/remove-link-to-parent.component.ts
